Change behavior of next-iteration
This commit changes the behavior of the next.iteration command compared
to the behavior in the core.itero module. In core.itero, the new heading
or list item is inserted immediately after the line the cursor is on.
The version in this commit instead inserts the new item after the end of
the node.

Addresses the issue discussed in
https://github.com/nvim-neorg/neorg/issues/938#issuecomment-1866310365.