Add promote and demote keybinds
New keybind commands to promote or demote the current heading or subtree
(that is, the current heading and its children). Note, here promoting a
heading means to move the heading closer to the root of the tree (that
is, to reduce the heading number), following the org-mode convention.
This is opposite to the convention used by Neorg's core.promo module.