Fix behavior of toggle-heading outside of heading
Previously, `toggle_heading()` would throw an error when trying to
create a heading on a line that does not fall within the heading tree,
as when making the first heading in a new document. This commit fixes
this issue; now, such a line is properly converted to a level-1 heading.